refactor: split refactor source helpers
